requirements.txt

pandas
openpyxl
ccxt
pyyaml
